+++ src/main/java/floud/demo/dto/friendship/FriendDeleteResponseDto.java
@@ -1,4 +1,4 @@
-package floud.demo.dto.mypage;
+package floud.demo.dto.friendship;
 
 import floud.demo.domain.enums.FriendshipStatus;
 import lombok.Builder;
@@ -6,8 +6,9 @@ import lombok.Getter;
 
 @Getter
 @Builder
-public class MypageFriendUpdateRequestDto {
+public class FriendDeleteResponseDto {
     private Long friendship_id;
-    private String nickname;
+    private String to_user;
+    private String from_user;
     private FriendshipStatus friendshipStatus;
 }
